Added additional models and metrics
Added KFold and StratifiedKFold for the Logistic Regression and Linear Regression model, with rmse used to calculate error for linear model. Compiled the Decision Trees and Random Forest models to run, looking at ranked importance and potential ideas to reduce the number of features. Need to explore more of the hyperparameters adjustments.